Remove core.js
